1267 - add a seperate component for the stripe form
